+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - GATT Service: Broadcast Audio Scan Service Server and Client (BASS 1.0)
 - hci_dump_stdout: allow to truncate ACL, SCO and ISO packets with HCI_DUMP_STDOUT_MAX_SIZE_*
+- example/a2dp_source_demo: add command to scan and connect to Bluetooth speaker by class of device
 
 ### Fixed
 - ESP32: fix init for BR/EDR Only mode